app icon added, validated fields and confirm dialog implemented

+++ app/src/main/java/com/example/curso/FormActivity.kt
@@ -3,8 +3,10 @@ package com.example.curso
 import Data.DataDbHelper
 import Model.Contact
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.os.Message
 import android.util.Log
 import android.view.View
 import android.widget.ArrayAdapter
@@ -42,7 +44,8 @@ class FormActivity : AppCompatActivity() {
 
         val phoneTypes:  ArrayList<String> = ArrayList()
         phoneTypes.add("Home")
-        phoneTypes.add("Cell")
+        phoneTypes.add("Job")
+        phoneTypes.add("Personal")
         phoneTypes.add("Other")
         val adp: ArrayAdapter<String> = ArrayAdapter(this, android.R.layout.simple_spinner_dropdown_item, phoneTypes)
         sp = findViewById(R.id.spinnerPhoneType)
@@ -80,31 +83,67 @@ class FormActivity : AppCompatActivity() {
         val contact = Contact(editName.text.toString(), editPhone.text.toString(),
             editDate.text.toString(), sp.selectedItemId.toString())
         var result: Long = -1
+        val regex = Regex("(6|7)(\\d){8}")
 
         if(editName.text.isEmpty() || editPhone.text.isEmpty()) {
             Toast.makeText(this, "Please enter required fields.",Toast.LENGTH_SHORT).show()
         } else {
-            if(idContact>-1) { //Update
-                contact.setId(idContact)
-                result = db!!.updateContact(contact).toLong()
-            }
-            else { //Create
-                result = db!!.insertContact(contact)
-            }
+            if(editPhone.text.toString().matches(regex)) {
+                if(idContact>-1) { //Update
+                    val dialog = AlertDialog.Builder(this)
+                        .setTitle("Attention, updating contact!!!")
+                        .setMessage("Are you sure to update the contact?")
+                        .setNegativeButton("Cancel") { view, _ ->
+                            view.dismiss()
+                        }
+                        .setPositiveButton("Accept") { view, _ ->
+                            view.dismiss()
+                            contact.setId(idContact)
+                            result = db!!.updateContact(contact).toLong()
+                            if(result>-1){
+                                Toast.makeText(applicationContext, "Contact saved successfully.",Toast.LENGTH_SHORT).show()
+                            } else {
+                                Toast.makeText(applicationContext, "Contact not saved.",Toast.LENGTH_SHORT).show()
+                            }
+                            finish()
+                        }
+                        .setCancelable(false)
+                        .create()
+                    dialog.show()
+                }
+                else { //Create
+                    result = db!!.insertContact(contact)
+                    if(result>-1){
+                        Toast.makeText(applicationContext, "Contact saved successfully.",Toast.LENGTH_SHORT).show()
+                    } else {
+                        Toast.makeText(applicationContext, "Contact not saved.",Toast.LENGTH_SHORT).show()
+                    }
+                    finish()
+                }
 
-            if(result>-1){
-                Toast.makeText(applicationContext, "Contact saved successfully.",Toast.LENGTH_SHORT).show()
-            } else {
-                Toast.makeText(applicationContext, "Contact not saved.",Toast.LENGTH_SHORT).show()
             }
-            finish()
+            else {
+                Toast.makeText(applicationContext, "Invalid phone number.",Toast.LENGTH_SHORT).show()
+            }
         }
     }
 
     fun removeContact(view: View) {
         db= DataDbHelper(this)
-        db!!.deleteContact(idContact)
-        Toast.makeText(applicationContext, "Contact removed successfully.",Toast.LENGTH_SHORT).show()
-        finish()
+        val dialog = AlertDialog.Builder(this)
+            .setTitle("Attention, eliminating contact!!!")
+            .setMessage("Are you sure to delete the contact?")
+            .setNegativeButton("Cancel") { view, _ ->
+                view.dismiss()
+            }
+            .setPositiveButton("Accept") { view, _ ->
+                view.dismiss()
+                db!!.deleteContact(idContact)
+                Toast.makeText(applicationContext, "Contact removed successfully.",Toast.LENGTH_SHORT).show()
+                finish()
+            }
+            .setCancelable(false)
+            .create()
+        dialog.show()
     }
 }
